Support: Move url module to lib/url/support
