+++ client/my-sites/upgrades/domain-management/transfer/transfer-prohibited.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { translate } from 'lib/mixins/i18n';
 import Card from 'components/card';
 import SectionHeader from 'components/section-header';
-import support from 'lib/support';
+import support from 'lib/url/support';
 
 const TransferProhibited = () => (
 	<div>
